reactor: Move read_directory() to posix_file_impl
This method is a thread-pool wrapper over raw readdir64 syscall. It's
used only by posix_file_impl's list_directory() API, but this file impl
can talk to thread-pool on its own (all the more so it already does so
in many other ways).
